+++ src/DGtal/base/ModuleSRC.cmake
@@ -1,5 +1,5 @@
 
 set(DGTAL_SRC ${DGTAL_SRC} 
-    DGtal/base/Bits.cpp
-    DGtal/base/Common.cpp)
+    
+)
