+++ src/DGtal/io/ModuleSRC.cmake
@@ -6,13 +6,10 @@
 #### boards
 ##########################################
 
-#set(DGTAL_SRC ${DGTAL_SRC}
-#  DGtal/io/Color.cpp)
+#set(DGTAL_SRC ${DGTAL_SRC})
 
 
-set(DGTALIO_SRC ${DGTALIO_SRC}
-  DGtal/io/boards/Board2D.cpp
-  DGtal/io/boards/Board3D.cpp)
+#set(DGTALIO_SRC ${DGTALIO_SRC})
 
 
 if( WITH_CAIRO )
