--- src/DGtal/io/boards/Board3D.cpp
@@ -1,46 +1,0 @@
-/**
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Lesser General Public License as
- *  published by the Free Software Foundation, either version 3 of the
- *  License, or  (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; withoutOBJ even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- **/
-
-/**
- * @file   Board3D.cpp
- * @author Aline Martin
- * @date   vendredi 7 juin 2013
- *
- * @brief
- *
- * Class for OBJ export
- * This is the implementation of methods defined in Board3D.h
- *
- * This file is part of the DGtal library.
- */
-
-///////////////////////////////////////////////////////////////////////////////
-#include "DGtal/io/boards/Board3D.h"
-#include <limits>
-//#include <vector>
-
-
-
-
-///////////////////////////////////////////////////////////////////////////////
-
-using namespace std;
-using namespace DGtal;
-
-///////////////////////////////////////////////////////////////////////////////
-// class Board3D
-///////////////////////////////////////////////////////////////////////////////
-
