--- src/DGtal/io/boards/Board3DTo2D.cpp
@@ -1,29 +1,0 @@
-/**
-*  This program is free software: you can redistribute it and/or modify
-*  it under the terms of the GNU Lesser General Public License as
-*  published by the Free Software Foundation, either version 3 of the
-*  License, or  (at your option) any later version.
-*
-*  This program is distributed in the hope that it will be useful,
-*  but WITHOUT ANY WARRANTY; without even the implied warranty of
-*  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-*  GNU General Public License for more details.
-*
-*  You should have received a copy of the GNU General Public License
-*  along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*
-**/
-
-/**
-* @file   Board3DTo2D.cpp
-* @author Martial Tola <http://liris.cnrs.fr/martial.tola/>
-* @date   mercredi 22 juin 2011
-*
-* @brief
-*
-* Class for PDF, PNG, PS, EPS, SVG export drawings with Cairo with 3D->2D projection
-* This is the implementation of methods defined in Board3DTo2D.h
-*
-* This file is part of the DGtal library.
-*/
-
